requirements.txt
discord.py==2.3.2
mariadb==1.1.10
requests==2.32.3